add resolver interface and secret resolver implementation
